+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.22.0 (TBD)
 
 - Added Stereoscope, Scalding Jelly, and Flakseed
+- Added mastery for Racket Launcher
 - Fixed potential null deref caused by event handler for destroyed Allay companion
 
 ## 1.21.2 (2025-01-30)